Update Exercise for Markdown inclusion
